FIX COLOURS FROM XPM:
Fixing colours by getting the correct data from the addr ptr.
